fix comments mentioned in review
