+++ packages/cli/src/modules/agents/__tests__/agents-service-reconstruct-gating.test.ts
@@ -16,6 +16,7 @@ import type { WorkflowRunner } from '@/workflow-runner';
 import type { WorkflowFinderService } from '@/workflows/workflow-finder.service';
 
 import type { AgentExecutionService } from '../agent-execution.service';
+import type { AgentSkillsService } from '../agent-skills.service';
 import type { AgentsToolsService } from '../agents-tools.service';
 import { AgentsService } from '../agents.service';
 import type { Agent } from '../entities/agent.entity';
@@ -58,6 +59,7 @@ function makeService(agentsToolsService: AgentsToolsService): AgentsService {
 		mock<N8nMemory>(),
 		mock<AgentExecutionService>(),
 		mock<AgentPublishedVersionRepository>(),
+		mock<AgentSkillsService>(),
 	);
 }
