+++ packages/frontend/@n8n/i18n/src/locales/en.json
@@ -5785,6 +5785,7 @@
 	"agents.builder.skills.create.cancel": "Cancel",
 	"agents.builder.skills.create.save": "Save skill",
 	"agents.builder.skills.create.error": "Could not create skill",
+	"agents.builder.skills.saveError": "Could not save skill",
 	"agents.builder.skills.validation.nameRequired": "Enter a title.",
 	"agents.builder.skills.validation.nameMaxLength": "Title must be 128 characters or fewer.",
 	"agents.builder.skills.validation.descriptionRequired": "Enter a description.",
